requirements.txt
requests
sqlalchemy
python-dotenv